06.10.2019

+++ src/main/java/pl/coderslab/app/web/meetings/MeetingService.java
@@ -1,8 +1,8 @@
 package pl.coderslab.app.web.meetings;
 
 import org.springframework.stereotype.Service;
-import pl.coderslab.app.FinderFormDTO;
-import pl.coderslab.app.repositories.NotificationRepository;
+import pl.coderslab.app.web.EmailService;
+import pl.coderslab.app.web.FinderFormDTO;
 import pl.coderslab.app.web.meetings.CoordianteJsonStructure.Coordinate;
 import pl.coderslab.app.web.meetings.DistanceJsonStructure.Distance;
 import pl.coderslab.app.web.meetings.member.MemberDTO;
@@ -18,6 +18,7 @@ import pl.coderslab.app.web.user.notifications.NotificationService;
 
 import javax.json.bind.Jsonb;
 import javax.json.bind.JsonbBuilder;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -27,6 +28,7 @@ import java.net.URLConnection;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.Base64;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -39,26 +41,35 @@ public class MeetingService {
     private CommentRepository commentRepository;
     private UserRepository userRepository;
     private NotificationService notificationService;
+    private HttpServletRequest request;
+    private EmailService emailService;
 
-    public MeetingService(MeetingRepository meetingRepository, UserService userService, CommentRepository commentRepository, UserRepository userRepository, NotificationService notificationService) {
+    public MeetingService(HttpServletRequest request, EmailService emailService,MeetingRepository meetingRepository, UserService userService, CommentRepository commentRepository, UserRepository userRepository, NotificationService notificationService) {
         this.meetingRepository = meetingRepository;
         this.userService = userService;
         this.commentRepository = commentRepository;
         this.userRepository = userRepository;
         this.notificationService = notificationService;
+        this.emailService = emailService;
+        this.request = request;
     }
 
     public Meeting getMeetingById(Long id, boolean alldata) {
         Meeting result = meetingRepository.findOne(id);
-        if (alldata) {
-            result.getMembers().size();
-            result.setBase64fromOwnerAvatar();
-            result.setComments(commentRepository.getAllByParentAndMeeting(null,result));
-            // w zasadzie nieograniczona ilosc komentarzy do komentarzy
-            for (Comment comment : result.getComments()) {
-                comment.setBase64fromUserAvatar();
-                comment.setChildren(getChildrenLoop(comment.getChildren()));
+
+        if (result!=null) {
+            if (alldata) {
+                result.getMembers().size();
+                result.getMembers().forEach(m -> m.setBase64Image(Base64.getEncoder().encodeToString(m.getAvatar())));
+                result.setBase64fromOwnerAvatar();
+                result.setComments(commentRepository.getAllByParentAndMeeting(null, result));
+                // w zasadzie nieograniczona ilosc komentarzy do komentarzy
+                for (Comment comment : result.getComments()) {
+                    comment.setBase64fromUserAvatar();
+                    comment.setChildren(getChildrenLoop(comment.getChildren()));
+                }
             }
+            setAvailableUsers(result);
         }
         return result;
     }
@@ -125,10 +136,10 @@ public class MeetingService {
     private Coordinate getCoordinates(String address) {
         String googleURL = "https://maps.googleapis.com/maps/api/geocode/json?address="
                 +address.replaceAll(" ","+")
-                + "&key=AIzaSyC5EJjfoZUTXckzVuwbvm3Ke0SWYwoi6OI";
+                +"&key="+ request.getServletContext().getInitParameter("apiKey");
         Jsonb jsonb = JsonbBuilder.create();
-        Coordinate result = jsonb.fromJson(getJSONStringFromUrl(googleURL), Coordinate.class);
-
+        //brak internetu - strona sie wysypie.
+        Coordinate result =  jsonb.fromJson(getJSONStringFromUrl(googleURL), Coordinate.class);
         return result;
     }
 
@@ -137,8 +148,10 @@ public class MeetingService {
                 + finderFormDTO.getLatitude() + "," + finderFormDTO.getLongitude()
                 + "&destinations="
                 + destinationAddres.replaceAll(" ","+")
-                + "&key=AIzaSyC5EJjfoZUTXckzVuwbvm3Ke0SWYwoi6OI";
+                + "&key="+request.getServletContext().getInitParameter("apiKey");
+
         Jsonb jsonb = JsonbBuilder.create();
+        // brak internetu strona sie wysypie
         Distance distance = jsonb.fromJson(getJSONStringFromUrl(googleURL), Distance.class);
 
 
@@ -172,31 +185,58 @@ public class MeetingService {
 
     public void updateMeeting(MeetingEditDTO meetingDTO) {
         Meeting meeting = meetingRepository.findOne(meetingDTO.getId());
+        if (meeting != null) {
 
-        if (meetingDTO.getSendNotification()) {
-            // przygotowanie tresci powiadomienia
-            String notificationText = "";
-            if (!meetingDTO.getAddress().equals(meeting.getAddress())) {
-                notificationText += "adres ";
-            }
-            if (!meetingDTO.getMeetTime().equals(meeting.getMeetTime())) {
-                notificationText += "data i godzina spotkania ";
-            }
-            if (!meetingDTO.getDescription().equals(meeting.getDescription())) {
-                notificationText += "opis ";
-            }
-            if (!meetingDTO.getTitle().equals(meeting.getTitle())) {
-                notificationText += "tytuł";
-            }
-            if (notificationText.length() > 0) {
-                notificationText = "Nastąpila zmiana wydarzenia <strong>" +
-                        meeting.getTitle() + "</strong> : " + notificationText;
-                notificationService.addNotificationForUserList(notificationText, "meetings?id=" + meeting.getId(), getMeetingById(meetingDTO.getId(), true).getMembers(),"primary");
-                // powiadomienie zapisane
-            }
+            if (meetingDTO.getSendNotification() || meetingDTO.getSendEmail()) {
+                // przygotowanie tresci powiadomienia
+                String notificationText = "";
+                String emailText = "";
+                if (!meetingDTO.getAddress().equals(meeting.getAddress())) {
+                    notificationText += "adres ";
+                    emailText += "  Zmiana adresu spotkania : <strong>"+meetingDTO.getAddress()+"</strong><br>";
+                }
+                if (!meetingDTO.getMeetTime().equals(meeting.getMeetTime())) {
+                    notificationText += "data i godzina spotkania ";
+                    emailText += "  Zmiana godziny/daty spotkania : <strong>"+meetingDTO.getMeetTime().format(DateTimeFormatter.ofPattern("yyyy-MM-dd, HH:mm"))+"</strong><br>";
+                }
+                if (!meetingDTO.getDescription().equals(meeting.getDescription())) {
+                    notificationText += "opis ";
+                    emailText += "  Zmiana opisu : <strong>"+meetingDTO.getDescription()+"</strong><br>";
+                }
+                if (!meetingDTO.getTitle().equals(meeting.getTitle())) {
+                    notificationText += "tytuł";
+                    emailText += "  Zmiana tytułu wydarzenia : <strong>"+meetingDTO.getTitle()+"</strong><br>";
+                }
 
+                if (notificationText.length() > 0) {
+                    notificationText = "Nastąpila zmiana wydarzenia <strong>" +
+                            meeting.getTitle() + "</strong> : " + notificationText;
+
+                    emailText = "Nastąpila zmiana wydarzenia <strong>" +
+                            meeting.getTitle() + "</strong><br>" + emailText + "<br>Pozdrowienia Meetings :)";
+
+
+
+                    if (meetingDTO.getSendNotification()) {
+                        notificationService.addNotificationForUserList(notificationText, "meetings?id=" + meeting.getId(), getMeetingById(meetingDTO.getId(), true).getMembers(), "primary");
+                        // powiadomienie zapisane
+                    }
+
+                    if (meetingDTO.getSendEmail()) {
+                        for (User member : meeting.getMembers()) {
+                            try {
+                                emailService.sendEmail(member.getEmail(),"Zmiany w spotkaniu \""+meeting.getTitle()+"\"",emailText);
+                            } catch (Exception e) {
+                                e.printStackTrace();
+                            }
+                        }
+                        // email wyslany
+                    }
+
+                }
+            }
+            fillMeetingAndSave(meetingDTO, meeting);
         }
-        fillMeetingAndSave(meetingDTO, meeting);
     }
 
 
@@ -214,31 +254,41 @@ public class MeetingService {
     }
     public void toogleMemberInMeeting(MemberDTO memberDTO) {
         Meeting meeting = meetingRepository.findOne(memberDTO.getMeetingId());
-        User user = userRepository.findOne(memberDTO.getUserId());
-        meeting.getMembers().size();
+        if (meeting!=null) {
+            User user = userRepository.findOne(memberDTO.getUserId());
+            meeting.getMembers().size();
 
-        String notificationText = "";
-        String alertType = "";
-        if (meeting.getMembers().contains(user)) {
-            System.out.println("Usuwamy uzytkownika z listy członków");
-            notificationText="Uzytkownik <strong>"
-                    +user.getFullName()
-                    +"</strong> opuścił wydarzenie <strong>"
-                    +meeting.getTitle()+"</strong>";
-            alertType = "danger";
-            meeting.removeMember(user);
-        } else {
-            System.out.println("Dodajemy uzytkownika do listy członków");
-            notificationText="Uzytkownik <strong>"
-                    +user.getFullName()
-                    +"</strong> bierze udział w wydarzeniu <strong>"
-                    +meeting.getTitle()+"</strong>";
-            alertType = "success";
-            meeting.addMember(user);
+            String notificationText = "";
+            String alertType = "";
+            if (meeting.getMembers().contains(user)) {
+                System.out.println("Usuwamy uzytkownika z listy członków");
+                notificationText = "Uzytkownik <strong>"
+                        + user.getFullName()
+                        + "</strong> opuścił wydarzenie <strong>"
+                        + meeting.getTitle() + "</strong>";
+                alertType = "danger";
+                meeting.removeMember(user);
+            } else {
+                try {
+                    emailService.sendEmail(userService.getUserById(memberDTO.getUserId()).getEmail(),"Zaplanowane spotkanie \""
+                            +meeting.getTitle()+"\"", "Witamy na spotkaniu "+meeting.getTitle()+"<br />"+"Pamiętaj spotkanie odbędzie sie w dniu/godz "
+                            +meeting.getMeetTime().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"))+" w miejscu "
+                            +meeting.getAddress()+".<br />Poniżej opis spotkania do zobaczenia.<br />"+meeting.getDescription());
+                } catch (Exception e) {
+                    System.out.println(e.getMessage());
+                }
+                System.out.println("Dodajemy uzytkownika do listy członków");
+                notificationText = "Uzytkownik <strong>"
+                        + user.getFullName()
+                        + "</strong> bierze udział w wydarzeniu <strong>"
+                        + meeting.getTitle() + "</strong>";
+                alertType = "success";
+                meeting.addMember(user);
+            }
+
+            notificationService.addNotificationForUser(notificationText, "meetings?id=" + meeting.getId(), meeting.getOwner(), alertType);
+            meetingRepository.save(meeting);
         }
-
-        notificationService.addNotificationForUser(notificationText,"meetings?id=" + meeting.getId(), meeting.getOwner(),alertType);
-        meetingRepository.save(meeting);
     }
 
     public List<Meeting> getMeetingByOwner(User owner) {
@@ -262,8 +312,18 @@ public class MeetingService {
     }
 
     public void removeMeeting(Long meetingId) {
-        commentRepository.deleteAllByMeetingId(meetingId);
-        meetingRepository.delete(meetingId);
+        Meeting meeting = meetingRepository.findOne(meetingId);
+
+        if (meeting!= null) {
+            // powiadomienie
+            String notificationText = "Wydarzenie <strong>" + meetingRepository.findOne(meetingId).getTitle() + "</strong> którego jesteś członkiem zostało usunięte";
+            String alertType = "dark";
+            // uwaga nowosc kopiujemy obiekt nie podajemy tylko referencji
+            notificationService.addNotificationForUserList(notificationText, "", meetingRepository.findOne(meetingId).getMembers(), alertType);
+
+            commentRepository.deleteAllByMeetingId(meetingId);
+            meetingRepository.delete(meetingId);
+        }
     }
 
     public List<Meeting> getMeetingByOwnerContainsFraze(String searchFraze, User owner) {
@@ -280,13 +340,11 @@ public class MeetingService {
 
     private List<Meeting> filterMeetings(String searchFraze, List<Meeting> result) {
         result = result.stream()
-                .filter(m -> {
-                    return (
-                            m.getAddress().contains(searchFraze) ||
-                                    m.getTitle().contains(searchFraze) ||
-                                    m.getMeetTime().format(DateTimeFormatter.ofPattern("yyyy-MM-dd, HH:mm")).contains(searchFraze)
-                    );
-                })
+                .filter(m -> (
+                        m.getAddress().contains(searchFraze) ||
+                                m.getTitle().contains(searchFraze) ||
+                                m.getMeetTime().format(DateTimeFormatter.ofPattern("yyyy-MM-dd, HH:mm")).contains(searchFraze)
+                ))
                 .collect(Collectors.toList());
 
         result.forEach(m -> {
@@ -295,5 +353,13 @@ public class MeetingService {
             m.setBase64fromOwnerAvatar();
         });
         return result;
+    }
+
+    private void setAvailableUsers(Meeting meeting) {
+        List<User> availableUsers = userService.findAllLoggedInUsers();
+        if (availableUsers.contains(meeting.getOwner())) {
+            meeting.getOwner().setAvailable(true);
+        }
+        meeting.getMembers().forEach(m->{if (availableUsers.contains(m)) m.setAvailable(true);});
     }
 }
